% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_data.R
\docType{package}
\name{side_data}
\alias{side_data}
\alias{side_metadata}
\alias{side_data-package}
\title{SIDE Data Package}
\usage{
side_metadata(path = "package/extdata/side_metadata.csv")
}
\arguments{
\item{path}{Specify path. Default refers to csv file located in package directory.}
}
\description{
Facilitates access to the SIDE data.

Function to load meta data of the SIDE data set
}
\seealso{
\url{url-to-paper}; \url{https://github.com/carl-mc/side_data}, \url{icr.ethz.ch/side}
}
\author{
Carl Müller-Crepon and Philipp Hunziker
}
