% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_data.R
\name{side_download}
\alias{side_download}
\title{Download a set of SIDE maps}
\usage{
side_download(sideid = c(), mapid = c(), country = c(), year = c(),
  dhs.round = c(), dhs.subround = c(), groupname = c(), marker = c(),
  dest.dir = getwd(), overwrite = FALSE, conv.hull = FALSE)
}
\arguments{
\item{sideid}{Vector of sideid as listed in `side_metadata()`.}

\item{mapid}{Vector of mapid as listed in `side_metadata()`.}

\item{country}{Vector of country names as specified in `side_metadata()`}

\item{year}{Vector of country codes as specified in `side_metadata()`}

\item{dhs.round}{Vector of DHS rounds as specified in `side_metadata()`}

\item{dhs.subround}{Vector of DHS subrounds as specified in `side_metadata()`}

\item{groupname}{Vector of group names as specified in `side_metadata()`}

\item{marker}{Type of group; one of `c("ethnic","religion","ethnic.religion")`}

\item{dest.dir}{Destination directory. If `conv.hull = TRUE`, a subdirectory `dest.dir/conv_hull/` will be created in which raster files of the convex hulls are saved.}

\item{overwrite}{Overwrite existing SIDE Data}

\item{conv.hull}{Download convex hull of DHS points used for interpolation?}
}
\description{
Facilitates bulk-download of SIDE data from \url{icr.ethz.ch/side}.
}
\details{
Specify the subset of SIDE maps you want to download using the above parameters. 
The inner join of all parameters determines which maps are downloaded.
}
