% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_data.R
\name{side_load}
\alias{side_load}
\title{Open a set of SIDE maps}
\usage{
side_load(sideid = c(), mapid = NULL, country = NULL, year = NULL,
  dhs.round = NULL, dhs.subround = NULL, groupname = c(), marker = c(),
  source.dir = getwd(), conv.hull = FALSE,
  side.metadata.df = side_metadata())
}
\arguments{
\item{sideid}{Vector of sideid.}

\item{mapid}{Single mapid as listed in `side_metadata()`.}

\item{country}{Single country}

\item{year}{Single year}

\item{dhs.round}{Single DHS round}

\item{dhs.subround}{Single DHS subroud; needed if multiple DHS survey in the same country and survey wave}

\item{groupname}{Name of group}

\item{marker}{Type of group; one of `c("ethnic","religion","ethnic.religion")`}

\item{source.dir}{Directory in which SIDE data is stored}

\item{conv.hull}{Download convex hull of DHS points used for interpolation?}

\item{side.metadata.df}{Meta data set of SIDE, results from `side_metadata()`. You can speed up things a little bit, if you load that only once.}
}
\description{
Facilitates the loading of SIDE data.
}
\details{
Specify the subset of SIDE maps you want to load using the above parameters. 
To guarantee that maps form a raster-stack, parameters 
must specify a subset of maps that is based on a single DHS round from one country and year.
The inner join of all parameters determines which maps are loaded.
}
