% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_data.R
\docType{package}
\name{sidedata}
\alias{sidedata}
\alias{side_metadata}
\alias{sidedata-package}
\title{SIDE Data Package}
\usage{
side_metadata()
}
\description{
Facilitates access to the SIDE data.

Function to load meta data of the SIDE data set
}
\seealso{
\url{url-to-paper}; \url{https://github.com/carl-mc/sidedata}, \url{icr.ethz.ch/data/side}
}
\author{
Carl Müller-Crepon and Philipp Hunziker
}
