% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_data.R
\name{sidemap2data}
\alias{sidemap2data}
\title{Query SIDE meta-data matching a set of SIDE maps}
\usage{
sidemap2data(side.raster)
}
\arguments{
\item{side.raster}{any `raster*` object named with existing sideid(s)}
}
\value{
data.frame with the metadata for the bands of sider.raster.
}
\description{
Function to query SIDE metadata that matches a named SIDE raster
}
