% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_data.R
\name{sidemap2data}
\alias{sidemap2data}
\title{Query SIDE meta-data matching a set of SIDE maps}
\usage{
sidemap2data(side.raster, side.metadata.df = side_metadata())
}
\arguments{
\item{side.raster}{any `raster*` object named with existing sideid(s)}

\item{side.metadata.df}{Meta data set of SIDE, results from `side_metadata()`. 
You can speed up things a little bit, if you load that only once.}
}
\description{
Function to query SIDE metadata that matches a named SIDE raster
}
